Refactor on date verification and added momentjs as a date helper
