+++ package.json
@@ -13,6 +13,7 @@
     "dotenv": "^7.0.0",
     "express": "^4.16.4",
     "express-validator": "^5.3.1",
+    "moment": "^2.24.0",
     "mongoose": "^5.4.20",
     "ts-node": "^8.0.3",
     "typegoose": "^5.6.0",
